Abstracting D3D calls into LowLevelRenderer class.
- Also will be linking to commits in the README progress now.

+++ Sample/include/sample.h
@@ -7,5 +7,5 @@ public:
 
 	void update(float deltaTime, float runningTime);
 	void render();
-	void onResize(float width, float height);
+	void onResize(int width, int height);
 };
